Extracted listeners attaching into a separate function, so that it's possible to disable listeners in some unit tests.
It was causing a lot of extra toolbar.show / hide calls, which in turn required me to call toolbar.show.reset() in every TC. That was not an elegant solution, so instead I made it easier to disable.

What's more, I'm thinking about removing it completely from the constructor, and attach only on first show() method. That would be even more performance-friendly solution. And we'll be able to remove these _attachListeners stubs from init functions.